pyproject.toml
[tool.poetry]
name = "alphabet-game-plays-api"
version = "0.1.0"
description = ""
authors = ["Greg Finley <gregory.finley@gmail.com>"]

[tool.poetry.dependencies]
python = "3.11.1"
functions-framework = "^3.2.0"
mysqlclient = "^2.1.1"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
types-mysqlclient = "^2.1.5.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
